requirements.txt
Pillow==8.4.0
Pillow==9.5.0
streamlit==1.22.0
